requirements.txt
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.2.0
Pillow==9.5.0
scikit_image==0.19.3
scipy==1.9.1
thop==0.1.1.post2209072238
torch==1.11.0+cu113
torchmetrics==0.9.3
torchvision==0.12.0+cu113
tqdm==4.64.1
